Added Lots of Updates
I've added lots of updates that include making an `Adapter` that will simplify the process of adding new commands and interacting with the different options without the extra params needing to be passed. You can see in `delete.js` the usage of the adapter to simply delete a message and reply to it.

I've added a body parser to the web server and moved some of the options around.
An update on the `bot.js` file is the change in the adapter being added to the mix.

In the `updating_status.js` file you can see a new set of methods being used from the Adapter that is the addition of adding in routes to the web API endpoints! This allows for unauthenticated or authenticated routes based off of a secret key setup in the config files.